requirements.txt
streamlit
langchain
tiktoken
huggingface_hub
python-dotenv
transformers